iOS 8 fixes
List sort action picker did not work on iOS 8 ipad when trying to sort from ASAM list view.
List view delete on ipad was not being passed the ASAM array before init.  This was causing problems with the number of ASAMs title in the list view header.
Sort button on list view on phone was really small.  Changed from a segmented button to a toolbar bar to match the rest of the buttons.
Fixed a bounce in the header title when returning from an ASAM detail view to the ASAM list view on the ipad.
Navbar/header color was wrong in a couple places in ipad, settings popover, list popover, and asam map icon popovers.  Updated this to match the dark grey used across the app.
App crashed when clicking on 'Legal Information' in the about section on iPhone iOS8.  Fixed a problem with setting font on that table view.